pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "green-agent-template"
version = "0.1.0"
description = "A template for A2A green agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "httpx>=0.28.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
    "google-genai>=1.36.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "statsmodels>=0.14.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]

[project.scripts]
agentbeats-run = "agentbeats.run_scenario:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentbeats"]